Connection to header and idiotenseite deleted
The Connection to the header and idiotenseite has been deleted to make changes to the layout (e.g. geometry) or module related changes to the idiotenseite possible. The Advantages of the summary related Adaption outweights the quest for a high standardisation between the different summaries. Furthermore is it for beginners easier to download a repo without having to download the idiotenseite and header as well.